+++ backend/src/main/java/com/project2/domain/member/service/FollowingService.java
@@ -38,8 +38,7 @@ public class FollowingService {
 	}
 
 	@Transactional(readOnly = true)
-	public long getFollowingsCount(Long memberId) {
-		Member member = memberService.findByIdOrThrow(memberId);
+	public long getFollowingsCount(Member member) {
 		return followRepository.countByFollowing(member);
 	}
 }
